refactor: ScrollContainer, Time 컴포넌트로 분리
