+++ frontend/src/components/@common/ScrollContainer/styled.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const ScrollContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  height: 47rem;
+  overflow-y: scroll;
+  padding-right: 3rem;
+`;
+
+export default ScrollContainer;
